std: TcpListener::bind_one and bind_all
`TcpListener::bind` is not safe, because it quietly binds only to one
address. The common use case is to bind to both IPv4 and IPv6 addresses.

To address this issue, `TcpListener` gets two new functions: `bind_one`
and `bind_all`. `bind_one` is like `bind`, but fails, if parameter is
resolved to more than one address. `bind_all` binds to all addresses and
returns a vec of `TcpListener`.

Additionally, `TcpListener::bind_all` contains algorithm to bind to the
same port on all protocols if port is zero. It is not equivalent to
calling `bind_one` in loop, because in the latter case listeners will
most likely be bound to different ports.